[update] nifs-retrieve package namespaces
Each module belongs to "nifs.retrieve" namespace.
