+++ README.md
@@ -1,6 +1,6 @@
-NIFS Package
-============
+NIFS-retrieve Package
+======================
 
-A nifs package facilitates to handle the LHD experimental data from LABCOM, Kaiseki server, etc. for python users.
+A nifs-retrieve package facilitates to handle the LHD experimental data from LABCOM, Kaiseki server, etc. for python users.
 
 For more information, see the [documentation pages](https://nifs-software.github.io/nifs-retrieve/).
