+++ nifs/retrieve/database/_dbstore.cpp
