+++ nifs/retrieve/exceptions.py
