+++ nifs/retrieve/expseq.py
