+++ nifs/retrieve/rawdata/_retrieve_t.cpp
