+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.14
+numpy>=1.18
 xarray
 matplotlib
 psycopg2
